+++ react/store/MyReturnsPageAdd.tsx
@@ -269,7 +269,7 @@ class MyReturnsPageAdd extends Component<Props, State> {
   }
 
   selectOrder(order) {
-    this.prepareOrderData(order, this.state.settings, false)
+    this.prepareOrderData(order, this.state.settings, false, true)
     if (order.shippingData.address) {
       const complement =
         order.shippingData.address.complement !== null
@@ -307,7 +307,8 @@ class MyReturnsPageAdd extends Component<Props, State> {
   prepareOrderData = (
     order: any,
     settings: any,
-    updateEligibleOrders: boolean
+    updateEligibleOrders: boolean,
+    isSelectingOrder = false
   ) => {
     const thisOrder = order
 
@@ -449,8 +450,19 @@ class MyReturnsPageAdd extends Component<Props, State> {
         return { previousOrders, products }
       })
       .then(({ previousOrders, products }) => {
+        /**
+         * This check is prevent setting the products that will be returned
+         * during componentDidMount. Since this method is being called inside a
+         * loop, any order that takes longer to resolve in the initial setup will
+         * overwrite the products if user has already selected an order to view.
+         */
+        if (isSelectingOrder) {
+          this.setState({
+            orderProducts: products,
+          })
+        }
+
         this.setState({
-          orderProducts: products,
           eligibleOrders: previousOrders,
         })